Pull request

+++ src/components/Connections/ConnectionList/ConnectionServiceList.vue
@@ -2,9 +2,10 @@
   <div v-show="services.length > 0">
     {{ title }}
     <ul class="list">
-      <li class="list-el" v-for="service in services">
+      <li class="list-el" v-for="service in services" :key="service.service_id">
         {{ service.label }}
-        <div>
+        <div class="right-container">
+          <div class="valid-policy-dot" :class="[ policyColor(service) ]" />
           <el-button size="medium"
             @click="preview(service)">Preview</el-button>
           <el-button size="medium" type="primary"
@@ -37,6 +38,11 @@ export default {
         return message.topic == '/topic/verifiable-services/request-service-list/'
       })
     },
+    servicePolicyValidationMessages: function() {
+      return this.messages.filter(message => {
+        return message.topic == '/topic/verifiable-services/request-service-list/usage-policy/'
+      })
+    },
     acapyApiUrl: function() {
       return this.$session.get('acapyApiUrl')
     },
@@ -53,7 +59,24 @@ export default {
     serviceListMessages: {
       handler: function() {
         this.serviceListMessages.forEach(message => {
-          this.services = JSON.parse(message.content.services)
+          if (message.content.connection_id != this.connection.connection_id) {
+            return
+          }
+          this.services = message.content.services
+          this.delete_message(message.uuid)
+        })
+      },
+      deep: true
+    },
+    servicePolicyValidationMessages: {
+      handler: function() {
+        this.servicePolicyValidationMessages.forEach(message => {
+          var [service_id, policy_validation] = Object.entries(message.content)[0]
+          const serviceIndex = this.services.findIndex(s => s.service_id == service_id)
+          if (serviceIndex == -1) { return }
+          const service = this.services[serviceIndex]
+          service.policy_validation = JSON.parse(policy_validation)
+          this.$forceUpdate()
           this.delete_message(message.uuid)
         })
       },
@@ -144,6 +167,14 @@ export default {
       })
       return langBranches
     },
+    policyColor(service) {
+      const policy_validation = service.policy_validation
+      if (!policy_validation) { return }
+      if (policy_validation.code == 0) {
+        return 'green'
+      }
+      return 'red'
+    },
     async preview(service) {
       this.$emit('service-preview', {
         connection_id: this.connection.connection_id,
@@ -180,4 +211,26 @@ export default {
 .list-el:hover {
   background-color: #fafafa;
 }
+
+.right-container {
+  display: flex;
+  align-items: center;
+}
+
+.valid-policy-dot {
+  height: 15px;
+  width: 15px;
+  background-color: #bbbbbb;
+  border-radius: 50%;
+  margin-right: 10px;
+}
+
+.valid-policy-dot.green {
+  background-color: green;
+}
+
+.valid-policy-dot.red {
+  background-color: red;
+}
+
 </style>
